refactor: move AdminDashboard to its own file
Closes #140
